refactor: 유저 인증 관련 컴포넌트 공용 컴포넌트로 전환
