+++ app/_components/Auth/AuthFooter.tsx
